move tests, add tsconfig
